add cart and coupons

+++ src/main/java/com/lygin/beekeepersshopapi/repositories/UserRepository.java
@@ -1,5 +1,6 @@
 package com.lygin.beekeepersshopapi.repositories;
 
+import com.lygin.beekeepersshopapi.entity.Product;
 import com.lygin.beekeepersshopapi.entity.Role;
 import com.lygin.beekeepersshopapi.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 
     User findByRole(Role role);
